move numerics function into earthquake, add definitions for notes and actions
